split EmailModelFunctionalTest to 2, to avoid data provider hacking to test 2 different Mysql instances
